# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for the IoT Project API.",
        "title": "IoT Proecjt",
        "contact": {},
        "version": "1.0"
    },
    "host": "192.168.88.132:9090",
    "basePath": "/",
    "paths": {
        "/drivers": {
            "get": {
                "description": "Get all drivers from the database and returns them as JSON.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Driver"
                ],
                "summary": "Get all drivers",
                "responses": {
                    "200": {
                        "description": "List of drivers",
                        "schema": {
                            "$ref": "#/definitions/controller.DriverServerResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.DriverServerResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new driver",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Driver"
                ],
                "summary": "Create a new driver",
                "parameters": [
                    {
                        "description": "driver object that needs to be added",
                        "name": "driver",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Driver"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/drivers/{driverID}": {
            "get": {
                "description": "Retrieve a driver from the database by its ID and return it as JSON.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Driver"
                ],
                "summary": "Retrieve a driver by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Driver ID",
                        "name": "driverID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Driver information",
                        "schema": {
                            "$ref": "#/definitions/controller.DriverServerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.DriverServerResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controller.DriverServerResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.DriverServerResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update information of a driver based on the provided data in JSON format.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Driver"
                ],
                "summary": "Update a driver",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Driver ID",
                        "name": "driverID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Driver object to be updated",
                        "name": "driver",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Driver"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated driver information",
                        "schema": {
                            "$ref": "#/definitions/controller.DriverServerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.DriverServerResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.DriverServerResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a driver from the database based on its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Driver"
                ],
                "summary": "Delete a driver",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Driver ID",
                        "name": "driverID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Driver deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/controller.DriverServerResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.DriverServerResponse"
                        }
                    }
                }
            }
        },
        "/vehicles": {
            "get": {
                "description": "Retrieve all vehicles stored in the database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vehicle"
                ],
                "summary": "Retrieve all vehicles",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/controller.VehicleServerResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/controller.VehicleServerResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing vehicle with the provided information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vehicle"
                ],
                "summary": "Update a vehicle",
                "parameters": [
                    {
                        "description": "Vehicle information",
                        "name": "vehicle",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Vehicle"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/controller.VehicleServerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/controller.VehicleServerResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/controller.VehicleServerResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new vehicle with the provided information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vehicle"
                ],
                "summary": "Create a new vehicle",
                "parameters": [
                    {
                        "description": "Vehicle information",
                        "name": "vehicle",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Vehicle"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/controller.VehicleServerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/controller.VehicleServerResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/controller.VehicleServerResponse"
                        }
                    }
                }
            }
        },
        "/vehicles/{vehicleID}": {
            "get": {
                "description": "Retrieve a vehicle from the database by its unique identifier",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vehicle"
                ],
                "summary": "Retrieve a vehicle by its ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vehicle ID",
                        "name": "vehicleID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/controller.VehicleServerResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/controller.VehicleServerResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a vehicle from the database by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vehicle"
                ],
                "summary": "Delete a vehicle",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vehicle ID",
                        "name": "vehicleID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/controller.VehicleServerResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/controller.VehicleServerResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.DriverServerResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "status": {
                    "type": "string"
                }
            }
        },
        "controller.VehicleServerResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "status": {
                    "type": "string"
                }
            }
        },
        "model.Driver": {
            "type": "object",
            "required": [
                "driver_id",
                "driver_license"
            ],
            "properties": {
                "driver_id": {
                    "type": "integer"
                },
                "driver_license": {
                    "type": "string"
                },
                "driver_name": {
                    "type": "string"
                },
                "home_town": {
                    "type": "string"
                }
            }
        },
        "model.Vehicle": {
            "type": "object",
            "required": [
                "vehicle_id",
                "vehicle_plate"
            ],
            "properties": {
                "vehicle_id": {
                    "type": "integer"
                },
                "vehicle_name": {
                    "type": "string"
                },
                "vehicle_plate": {
                    "type": "string"
                }
            }
        }
    }
}